% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_sales_ts.R
\name{clean_sales_ts}
\alias{clean_sales_ts}
\title{Czyści dane szeregów czasowych}
\usage{
clean_sales_ts(data)
}
\arguments{
\item{data}{Ramka danych do wyczyszczenia}
}
\value{
Oczyszczona ramka danych bez braków w kolumnach numerycznych
}
\description{
Zastępuje braki danych (NA) w kolumnach numerycznych średnią wartością z danej kolumny.
}
\examples{
# clean_sales_ts(moje_dane)
}
