% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_sales_metrics.R
\name{compute_sales_metrics}
\alias{compute_sales_metrics}
\title{Oblicza kluczowe metryki biznesowe dla sprzedaży}
\usage{
compute_sales_metrics(data)
}
\arguments{
\item{data}{Oczyszczona ramka danych (np. wynik clean_sales_ts)}
}
\value{
Ramka danych (tibble) z jednym wierszem podsumowującym metryki
}
\description{
Funkcja wylicza całkowitą sprzedaż, średnią, zmienność (odchylenie standardowe)
oraz średnią liczbę produktów na promocji.
}
\examples{
# compute_sales_metrics(czyste_dane)
}
