% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_management_summary.R
\name{create_management_summary}
\alias{create_management_summary}
\title{Generuje podsumowanie biznesowe dla zarządu}
\usage{
create_management_summary(data)
}
\arguments{
\item{data}{Oczyszczona ramka danych (np. wynik clean_sales_ts)}
}
\value{
Lista zawierająca kluczowe wnioski biznesowe
}
\description{
Oblicza kluczowe wskaźniki takie jak najlepszy i najgorszy sklep pod względem
całkowitej sprzedaży oraz średnią sprzedaż dla całej sieci.
}
\examples{
# create_management_summary(czyste_dane)
}
