% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_prognosis.R
\name{create_prognosis}
\alias{create_prognosis}
\title{Generuje prognozę sprzedaży (ARIMA i Prophet)}
\usage{
create_prognosis(data, specific_store = NULL, horizon = 30)
}
\arguments{
\item{data}{Ramka danych ze sprzedażą}

\item{specific_store}{Opcjonalny numer sklepu (store_nbr) do analizy}

\item{horizon}{Horyzont prognozy (domyślnie 30 dni)}
}
\value{
Lista zawierająca modele ARIMA i Prophet (lub ich prognozy)
}
\description{
Domyślnie agreguje dane dla całej firmy, aby przyspieszyć obliczenia.
Użytkownik może opcjonalnie podać numer konkretnego sklepu.
}
\examples{
# prognoza <- create_prognosis(dane, specific_store = 1)
}
