% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_prognosis_ets.R
\name{create_prognosis_ets}
\alias{create_prognosis_ets}
\title{Create sales forecast using ETS}
\usage{
create_prognosis_ets(data, periods = 30)
}
\arguments{
\item{data}{Sales dataset}

\item{periods}{Number of future periods}
}
\value{
Forecast object
}
\description{
Create sales forecast using ETS
}
