% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_sales_data.R
\name{load_sales_data}
\alias{load_sales_data}
\title{Wczytuje i łączy dane sprzedażowe Kaggle}
\usage{
load_sales_data(train_path, stores_path, holidays_path)
}
\arguments{
\item{train_path}{Ścieżka do pliku train.csv}

\item{stores_path}{Ścieżka do pliku stores.csv}

\item{holidays_path}{Ścieżka do pliku holidays_events.csv}
}
\value{
Połączona ramka danych (tibble)
}
\description{
Funkcja wczytuje pliki train.csv, stores.csv oraz holidays_events.csv,
a następnie łączy je w jedną ramkę danych.
}
\examples{
# load_sales_data("train.csv", "stores.csv", "holidays_events.csv")
}
