% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_promotion_impact.R
\name{plot_promotion_impact}
\alias{plot_promotion_impact}
\title{Plot promotion impact}
\usage{
plot_promotion_impact(data)
}
\arguments{
\item{data}{Sales dataset}
}
\value{
ggplot object
}
\description{
Plot promotion impact
}
