% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sales_trends.R
\name{plot_sales_trends}
\alias{plot_sales_trends}
\title{Tworzy wizualizację trendów sprzedażowych}
\usage{
plot_sales_trends(data)
}
\arguments{
\item{data}{Ramka danych zawierająca kolumny 'date' i 'sales'}
}
\value{
Obiekt wykresu ggplot
}
\description{
Rysuje wykres liniowy pokazujący sprzedaż w czasie. Aby wykres był czytelny,
zaleca się wcześniejsze zagregowanie danych (np. do poziomu dziennego lub tygodniowego).
}
\examples{
# plot_sales_trends(czyste_dane)
}
