% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sales_ts_logic.R
\name{sales_ts_logic}
\alias{sales_ts_logic}
\title{Kompleksowa analiza sprzedaży dla wybranego sklepu}
\usage{
sales_ts_logic(data, filter_city = NULL, filter_type = NULL)
}
\arguments{
\item{data}{Ramka danych z połączonymi informacjami o sprzedaży i sklepach}

\item{filter_city}{Opcjonalna nazwa miasta do przefiltrowania (ciąg znaków)}

\item{filter_type}{Opcjonalny typ sklepu do przefiltrowania (ciąg znaków)}
}
\value{
Lista zawierająca dwa elementy: 'metryki' (data.frame) oraz 'wykres' (ggplot)
}
\description{
Funkcja wyższego rzędu: filtruje dane dla wybranego miasta i/lub typu sklepu,
a następnie oblicza metryki i generuje wykres.
}
\examples{
# wyniki <- sales_ts_logic(dane, filter_city = "Quito")
# wyniki$metryki
# print(wyniki$wykres)
}
