% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_sales_ts.R
\name{validate_sales_ts}
\alias{validate_sales_ts}
\title{Sprawdza jakość danych szeregów czasowych}
\usage{
validate_sales_ts(data)
}
\arguments{
\item{data}{Ramka danych do walidacji}
}
\value{
Lista z wynikami weryfikacji (wartości logiczne)
}
\description{
Weryfikuje ramkę danych pod kątem braków wartości (NA) i zduplikowanych wierszy.
}
\examples{
# validate_sales_ts(moje_dane)
}
