% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_sales_ts.R
\name{clean_sales_ts}
\alias{clean_sales_ts}
\title{Clean sales time series data}
\usage{
clean_sales_ts(data)
}
\arguments{
\item{data}{Sales dataset}
}
\value{
Cleaned dataset
}
\description{
Clean sales time series data
}
