% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_sales_metrics.R
\name{compute_sales_metrics}
\alias{compute_sales_metrics}
\title{Compute sales metrics}
\usage{
compute_sales_metrics(data)
}
\arguments{
\item{data}{Sales dataset}
}
\value{
Data frame with business metrics
}
\description{
Compute sales metrics
}
