% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_management_summary.R
\name{create_management_summary}
\alias{create_management_summary}
\title{Create management summary}
\usage{
create_management_summary(data)
}
\arguments{
\item{data}{Sales dataset}
}
\value{
Data frame with management summary
}
\description{
Create management summary
}
