% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_prognosis.R
\name{create_prognosis}
\alias{create_prognosis}
\title{Create sales forecast}
\usage{
create_prognosis(data, periods = 30)
}
\arguments{
\item{data}{Sales dataset}

\item{periods}{Number of future periods}
}
\value{
Forecast object
}
\description{
Create sales forecast
}
