% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_sales_data.R
\name{load_sales_data}
\alias{load_sales_data}
\title{Load sales data}
\usage{
load_sales_data(train_path, stores_path, holidays_path)
}
\arguments{
\item{train_path}{Path to train.csv}

\item{stores_path}{Path to stores.csv}

\item{holidays_path}{Path to holidays_events.csv}
}
\value{
Tibble
}
\description{
Load sales data
}
