% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sales_trends.R
\name{plot_sales_trends}
\alias{plot_sales_trends}
\title{Plot sales trends}
\usage{
plot_sales_trends(data)
}
\arguments{
\item{data}{Sales dataset}
}
\value{
ggplot object
}
\description{
Plot sales trends
}
