% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sales_ts_logic.R
\name{sales_ts_logic}
\alias{sales_ts_logic}
\title{Sales time series logic}
\usage{
sales_ts_logic(
  data,
  city = NULL,
  state = NULL,
  type = NULL,
  action = "metrics"
)
}
\arguments{
\item{data}{Sales dataset}

\item{city}{Optional city filter}

\item{state}{Optional state filter}

\item{type}{Optional store type filter}

\item{action}{"metrics" or "plot"}
}
\value{
Metrics table or plot
}
\description{
Sales time series logic
}
