% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_sales_ts.R
\name{validate_sales_ts}
\alias{validate_sales_ts}
\title{Validate sales time series data}
\usage{
validate_sales_ts(data)
}
\arguments{
\item{data}{Sales dataset}
}
\value{
List containing validation results
}
\description{
Validate sales time series data
}
